+++ CHANGELOG.md
@@ -11,7 +11,8 @@ All notable changes to this project will be documented in this file. The format 
 ### Backend
 - Fixed:
   -  Poor TypeORM performance in `/applications` endpoint [issue #1131](https://github.com/bloom-housing/bloom/issues/1131)
-  -  Add "ohaFormat" to CSV exporter
+- Added:
+  -  Added "ohaFormat" to CSV exporter (includes OHA and HOPWA preferences)
 
 ### Frontend